refactor: re-use existing eigentuemerMapping
