--- app/services/pdf/beratungshilfe/sections/F_besitz/util.ts
@@ -1,6 +1,0 @@
-export const eigentuemerMapping = {
-  myself: "Ich alleine",
-  partner: "Ehe-Partner:in",
-  myselfAndPartner: "Mein:e Ehe-Partner:in und ich gemeinsam",
-  myselfAndSomeoneElse: "Ich gemeinsam mit jemand anderem",
-} as const;
